Adds a ResponseEntity for the get() methods.
The reason is that the resource is identified in the URI, so if it
doesn't exist, the API should inform that with a 404.
The case with the list() method is different because the resource always
exists, but may be empty. So in this case a 200 with an empty list is
correct.