Adds a custom SDJ Repository method.
Together with Query Methods, @Query and orm.xml we now have four methods of implementing repository queries and this is the first one that uses Java code.
This is a dynamic method that handles any parameter related to Restaurant in a new endpoint (/beta/restaurants/find). This implementation uses pure JPQL and IFs to determine when to use something as a criterion.

RestaurantRepositoryImpl.java is a special class that Spring Data JPA (SDJ) detects and uses. This detection is based on the class name, which is the name of the repository interface plus the string "Impl". This is all that's necessary for SDJ to detect and use the methods implemented here, but, for the sake of code readability, we make this implementation implement an interface and we extend the interface on the default SDJ repository.