Adds Flyway to handle database migrations.
Adds the first script to create the database and import data (since it's a demo project, I'm ok with that).
The script is compatible with MySQL and H2 (given that you use the new datasource.url for h2 provided in this commit, which configures MySQL syntax for H2). Link to H2 documentation: http://h2database.com/html/features.html#compatibility.
Also fixes two inserts in import.sql (it's not used when Flyway is configured, but I'll keep it for now).