Introduces @ResponseStatus class annotation.
Up to this point we have been using this annotation on methods. Now we introduced this on a class, which happens to be an exception, so when it's thrown (not handled) the HTTP response code is defined by this annotation on the class. So exceptions are not handled on the controller anymore (for the example here, which is deletion of Restaurants).
This introduces something BAD: we're dealing with HTTP stuff on the service layer. That's no good, that's bad! Bad developer, bad!