Introduces Spring's ResponseStatusException class.
ResponseStatusException class allows you throw an exception with a specific message and HTTP status code. A good fit for simple PoCs.
With that, it's not necessary to handle HTTP codes on the service exception classes (ee how the exception classes are back to being basic), but you still are kind of limited.
Let's follow along...