Makes business exceptions extend ResponseStatusException.
In this approach, our exceptions extend ResponseStatusExceptions, which makes the Controller code much simpler, since it doesn't have to catch business exceptions and throw a new web exception. Also, each exception is overloaded with a constructor that adds the ability to customize the HTTP status code for the exception. It sounds awkward to e able to customize the HTTP status code in this project... because it is! We are not going to use it here. This would, though, be extremely useful if you want to create a unique and generic exception class (on small projects and PoCs) to return multiple HTTP status codes. When throwing you would add the corresponding HTTP status code for different business rules (conflict, not found etc.).

Ok, you notice that there's no way to fully segregate business and web layer. Yeah, I know. We are going to deal with this later.