Returns to @ResponseStatus class annotation.
Since we want specific business layer exceptions, we are returning to the solution where the business exception is annotated with the HTTP status code related to what the exception means. This couples api/http and business layers just a tiny little bit (the actual @ResponseStatus class annotation on the exceptions).
But combining this with a method on the service layer that, in this cases, finds an entity or else throws the corresponding business exception, we were able to make the controller code much more readable and simple.