Adds a GenericBusinessException to normalize API exception format.
City and Restaurant, which have related entities, were catching the problem of the related entity not existing and returning a custom ResponseEntity with an error. The problem is that this return is different from the business exceptions at this point, that were already using a format like so:

```json
{
  "timestamp": "2020-01-18T12:32:40.857+0000",
  "status": 404,
  "error": "Not Found",
  "message": "There's no Restaurant with the id 50.",
  "trace": "<omitted for brevity>",
  "path": "/restaurants/50"
}
```
By adding a GenericBusinessException, we can now control on the Controller the type of Exception to return according to the desired HTTP status code required. It's ok to have this logic on the controller: http-related stuff should be dealt with on the controller!