Adds better exception granularity.
Creates new custom ProvinceNotFound exception extending EntityNotFoundException. This is useful in cases where granular tratment/catch is necessary but you don't need to create exceptions for each and every entity. One must use its brain to decide which entities will have custom exceptions.
These new exceptions include a constructor that gets the entity ID and already formats the message, removing the need to format the messages on the caller classes when the exception is so specific already.
Throws that specific exceptions when dealing with something that throws them (in the Restaurant and City controllers, throwing Cuisine and Province exceptions, respectively).
Adds an option to throw GenericBusinessException with a cause (another exception that literally caused the generic one). Throws with cause on the controllers now.
Makes EntityNotFoundException extends GenericBusinessException, creating a more concise hierarchy.
Replicates this logic for City, Cuisine, Permission, PaymentMethod and Restaurant and then makes EntityNotFoundException abstract. This makes it still possible to catch this generic EntityNotFoundException (like we do in the controllers, in case there's more than one dependant entity), but we can not instantiate and throw, because we should throw the specific exceptions. This makes necessary to have specific exceptions for each entity. That's a decision. That's on if you believe that's the case.