Adds handlePropertyBindingException method.
This handles Jackson failures on deserialization.
Also makes the PATCH methods use the same JSON properties, catches IllegalArgumentException and throws a new HttpMessageNotReadableException. This makes possible for the API to ignore properties that exist but are @JsonIgnore(ed).

Note: only the Restaurant PATCH got this new code, for reference. Others are the same at this point.