Adds Validation Groups.
Now we can specify to which validation group (an interface in this case) each valiadation belongs and when validating we specify which group of validations we want to perform.
With this it's possible to specify, for example, that on Restaurant operations, the Cuisine ID must not be null, but on Cuisine operations, the Cuisine name must not be blank. By specifying groups, those validations are grouped in different groups.

Note: to use this we change the validation on the controllers from `@Valid` (from javax.validation) to `@Validated` (from org.springframework.validation.annotation).