Adds a Mixin class to solve a problem of resource representation.
This example is similar to what we did before with Restaurant and Cuisine representations. We want to be able to see the Province name and abbreviation when returning a City, but we do not want to allow the client to send Province name and abbreviation when creating a City. Without this new code those properties, since they are valid properties on the Province object, were ignored, but this is not the correct behaviour.

This new code combines a Mixin class and the `@JsonIgnoreProperties` annotation to fix the issue.