Adds DTOs and Model Mapper.
The RestaurantInputDisassembler.java and RestaurantModelAssembler.java use Model Mapper to convert from Input to Domain and Domain to Model, respectively.
RestaurantController.java is changed accordingly.
The api.model package holds the types of the API (both the Model and Input types).
The CuisineIdInput is a special type to have a cuisine with only an ID property.
The RestaurantMixin.java is removed since is no longer necessary. We use another strategy now.
There's also a configuration to provide the ModelMapper as a Spring bean for the app to autowire whenever necessary.