Adds the DTO layer for CityController.
In this case I only created the disassembler because it needs a special treatment for the Province when copying data. See code comments.

I don't like this solution, is not a good one for the developer. When will he use the Generic versus the specific disassembler? How to know that? Based on having a nested class? Not a good solution! =(