Adds a business rule to enforce the uniqueness of email on the user entity.
Notice the detach method to make sure that the persistence manager  doesn't persist the modifications to the User variable before the findByEmail method. This is an expected behviour from JPA and should be considered when developing these types of rules.

Alse fixes the update method of the UserCrudService to use the service save method with the business rule instead of going straight to the repository.