Adds two endpoints to open and close a restaurant.
This was not implemented as the active endpoint, which uses HTTP verbs to specify the semantics of the operations, but rather uses the URI to do just that with the same PUT HTTP operation on both endpoints (/open and /close).

NOTE: the fact that the endpoint is a verb is being validated but I'll keep it like that for now.