Adds an option to reduce entity fields to return with JsonFilter annotation.
JsonFilter is more powerful than JsonView since it allows us to filter programmatically for specific fields of the entity to be serialized.
In this example the field selection is even exposed to the API client to determine the fields to show via a param called fields with comma separated values.