Makes orders endpoint pageable.
Follows the same implementation from the previous commit with an important difference: it was necessary to consider the second query that Spring Data does when the request is paged, which is a count.

So, the OrderSpecs - that forms the predicates to filter the results based on Params sent by the client - needs to consider that the count query will also use the Spec, so when the result type is not a Order, we ignore all the fetchs since they would cause an exception.