Adds ETags (Entity Tags) and its verification.
ETags are a HTTP feature implemented by the browsers that returns a hash of the response to the client. When the response is staled (by cache-control header), the client can send the header if-none-match with the hash. Now, if the hash that the client has and the hash of the response matches, Spring returns a 304 (not modified), which makes the client make use of the staled version on its cache (in fact revalidating it).